Refactor update tests to separate files
