+++ quickwit/quickwit-integration-tests/src/tests/update_tests/doc_mapping_tests.rs
@@ -21,143 +21,13 @@ use std::collections::HashSet;
 use std::time::Duration;
 
 use quickwit_config::service::QuickwitService;
-use quickwit_rest_client::rest_client::CommitType;
-use quickwit_serve::SearchRequestQueryString;
 use serde_json::{json, Value};
 
-use crate::ingest_json;
-use crate::test_utils::{ingest_with_retry, ClusterSandbox};
+use super::assert_hit_count;
+use crate::test_utils::ClusterSandbox;
 
-async fn assert_hit_count(
-    sandbox: &ClusterSandbox,
-    index_id: &str,
-    query: &str,
-    expected_hits: Result<u64, ()>,
-) {
-    let search_res = sandbox
-        .searcher_rest_client
-        .search(
-            index_id,
-            SearchRequestQueryString {
-                query: query.to_string(),
-                ..Default::default()
-            },
-        )
-        .await;
-    if let Ok(expected_hit_count) = expected_hits {
-        let resp = search_res.unwrap_or_else(|_| panic!("query: {}", query));
-        assert_eq!(resp.errors.len(), 0, "query: {}", query);
-        assert_eq!(resp.num_hits, expected_hit_count, "query: {}", query);
-    } else if let Ok(search_response) = search_res {
-        assert!(!search_response.errors.is_empty(), "query: {}", query);
-    }
-}
-
-#[tokio::test]
-async fn test_update_search_settings_on_multi_nodes_cluster() {
-    quickwit_common::setup_logging_for_tests();
-    let nodes_services = vec![
-        HashSet::from_iter([QuickwitService::Searcher]),
-        HashSet::from_iter([QuickwitService::Metastore]),
-        HashSet::from_iter([QuickwitService::Indexer]),
-        HashSet::from_iter([QuickwitService::ControlPlane]),
-        HashSet::from_iter([QuickwitService::Janitor]),
-    ];
-    let sandbox = ClusterSandbox::start_cluster_nodes(&nodes_services)
-        .await
-        .unwrap();
-    sandbox.wait_for_cluster_num_ready_nodes(5).await.unwrap();
-
-    {
-        // Wait for indexer to fully start.
-        // The starting time is a bit long for a cluster.
-        tokio::time::sleep(Duration::from_secs(3)).await;
-        let indexing_service_counters = sandbox
-            .indexer_rest_client
-            .node_stats()
-            .indexing()
-            .await
-            .unwrap();
-        assert_eq!(indexing_service_counters.num_running_pipelines, 0);
-    }
-
-    // Create index
-    sandbox
-        .indexer_rest_client
-        .indexes()
-        .create(
-            r#"
-            version: 0.8
-            index_id: my-updatable-index
-            doc_mapping:
-              field_mappings:
-              - name: title
-                type: text
-              - name: body
-                type: text
-            indexing_settings:
-              commit_timeout_secs: 1
-            search_settings:
-              default_search_fields: [title]
-            "#,
-            quickwit_config::ConfigFormat::Yaml,
-            false,
-        )
-        .await
-        .unwrap();
-    assert!(sandbox
-        .indexer_rest_client
-        .node_health()
-        .is_live()
-        .await
-        .unwrap());
-
-    // Wait until indexing pipelines are started.
-    sandbox.wait_for_indexing_pipelines(1).await.unwrap();
-
-    // Check that ingest request send to searcher is forwarded to indexer and thus indexed.
-    ingest_with_retry(
-        &sandbox.searcher_rest_client,
-        "my-updatable-index",
-        ingest_json!({"title": "first", "body": "first record"}),
-        CommitType::Auto,
-    )
-    .await
-    .unwrap();
-    // Wait until split is committed and search.
-    tokio::time::sleep(Duration::from_secs(4)).await;
-    // No hit because default_search_fields covers "title" only
-    assert_hit_count(&sandbox, "my-updatable-index", "record", Ok(0)).await;
-    // Update index to also search "body" by default, search should now have 1 hit
-    sandbox
-        .searcher_rest_client
-        .indexes()
-        .update(
-            "my-updatable-index",
-            r#"
-            version: 0.8
-            index_id: my-updatable-index
-            doc_mapping:
-              field_mappings:
-              - name: title
-                type: text
-              - name: body
-                type: text
-            indexing_settings:
-              commit_timeout_secs: 1
-            search_settings:
-              default_search_fields: [title, body]
-            "#,
-            quickwit_config::ConfigFormat::Yaml,
-        )
-        .await
-        .unwrap();
-    assert_hit_count(&sandbox, "my-updatable-index", "record", Ok(1)).await;
-    sandbox.shutdown().await.unwrap();
-}
-
-/// Update the doc mapping between 2 local ingestion (separate indexing pipelines) and assert the
-/// number of hits for the given query
+/// Update the doc mapping between 2 calls to local-ingest (forces separate indexing pipelines) and
+/// assert the number of hits for the given query
 async fn validate_search_across_doc_mapping_updates(
     index_id: &str,
     original_doc_mapping: Value,
@@ -167,17 +37,17 @@ async fn validate_search_across_doc_mapping_updates(
     query_and_expect: &[(&str, Result<u64, ()>)],
 ) {
     quickwit_common::setup_logging_for_tests();
-    let nodes_services = vec![
-        HashSet::from_iter([QuickwitService::Searcher]),
-        HashSet::from_iter([QuickwitService::Metastore]),
-        HashSet::from_iter([QuickwitService::Indexer]),
-        HashSet::from_iter([QuickwitService::ControlPlane]),
-        HashSet::from_iter([QuickwitService::Janitor]),
-    ];
+    let nodes_services = vec![HashSet::from_iter([
+        QuickwitService::Searcher,
+        QuickwitService::Metastore,
+        QuickwitService::Indexer,
+        QuickwitService::ControlPlane,
+        QuickwitService::Janitor,
+    ])];
     let sandbox = ClusterSandbox::start_cluster_nodes(&nodes_services)
         .await
         .unwrap();
-    sandbox.wait_for_cluster_num_ready_nodes(5).await.unwrap();
+    sandbox.wait_for_cluster_num_ready_nodes(1).await.unwrap();
 
     {
         // Wait for indexer to fully start.
@@ -261,7 +131,7 @@ async fn validate_search_across_doc_mapping_updates(
 }
 
 #[tokio::test]
-async fn test_update_doc_mappings_text_to_u64() {
+async fn test_update_doc_mapping_text_to_u64() {
     let index_id = "update-text-to-u64";
     let original_doc_mappings = json!({
         "field_mappings": [
@@ -293,18 +163,20 @@ async fn test_update_doc_mappings_text_to_u64() {
 }
 
 #[tokio::test]
-async fn test_update_doc_mappings_u64_to_text() {
+async fn test_update_doc_mapping_u64_to_text() {
     let index_id = "update-u64-to-text";
     let original_doc_mappings = json!({
         "field_mappings": [
             {"name": "body", "type": "u64"}
-        ]
+        ],
+        "mode": "strict",
     });
     let ingest_before_update = &[json!({"body": 14}), json!({"body": 15})];
     let updated_doc_mappings = json!({
         "field_mappings": [
-            {"name": "body", "type": "text"}
-        ]
+            {"name": "body", "type": "text"},
+        ],
+        "mode": "strict",
     });
     let ingest_after_update = &[json!({"body": "16"}), json!({"body": "hello world"})];
     validate_search_across_doc_mapping_updates(
@@ -323,7 +195,7 @@ async fn test_update_doc_mappings_u64_to_text() {
 }
 
 #[tokio::test]
-async fn test_update_doc_mappings_json_to_text() {
+async fn test_update_doc_mapping_json_to_text() {
     let index_id = "update-json-to-text";
     let original_doc_mappings = json!({
         "field_mappings": [
@@ -357,7 +229,7 @@ async fn test_update_doc_mappings_json_to_text() {
 }
 
 #[tokio::test]
-async fn test_update_doc_mappings_json_to_object() {
+async fn test_update_doc_mapping_json_to_object() {
     let index_id = "update-json-to-object";
     let original_doc_mappings = json!({
         "field_mappings": [
@@ -398,7 +270,7 @@ async fn test_update_doc_mappings_json_to_object() {
 // TODO expected to be fix as part of #5084
 #[tokio::test]
 #[ignore]
-async fn test_update_doc_mappings_tokenizer_default_to_raw() {
+async fn test_update_doc_mapping_tokenizer_default_to_raw() {
     let index_id = "update-tokenizer-default-to-raw";
     let original_doc_mappings = json!({
         "field_mappings": [
@@ -437,7 +309,7 @@ async fn test_update_doc_mappings_tokenizer_default_to_raw() {
 // TODO expected to be fix as part of #5084
 #[tokio::test]
 #[ignore]
-async fn test_update_doc_mappings_tokenizer_add_position() {
+async fn test_update_doc_mapping_tokenizer_add_position() {
     let index_id = "update-tokenizer-add-position";
     let original_doc_mappings = json!({
         "field_mappings": [
@@ -473,7 +345,7 @@ async fn test_update_doc_mappings_tokenizer_add_position() {
 }
 
 #[tokio::test]
-async fn test_update_doc_mappings_tokenizer_raw_to_phrase() {
+async fn test_update_doc_mapping_tokenizer_raw_to_phrase() {
     let index_id = "update-tokenizer-raw-to-phrase";
     let original_doc_mappings = json!({
         "field_mappings": [
@@ -509,8 +381,37 @@ async fn test_update_doc_mappings_tokenizer_raw_to_phrase() {
 }
 
 #[tokio::test]
-async fn test_update_doc_mappings_add_on_dynamic() {
-    let index_id = "update-add-on-dynamic";
+async fn test_update_doc_mapping_strict_to_dynamic() {
+    let index_id = "update-strict-to-dynamic";
+    let original_doc_mappings = json!({
+        "field_mappings": [
+            {"name": "body", "type": "text"}
+        ],
+        "mode": "strict",
+    });
+    let ingest_before_update = &[json!({"body": "hello"})];
+    let updated_doc_mappings = json!({
+        "mode": "dynamic",
+    });
+    let ingest_after_update = &[json!({"body": "world", "title": "salutations"})];
+    validate_search_across_doc_mapping_updates(
+        index_id,
+        original_doc_mappings,
+        ingest_before_update,
+        updated_doc_mappings,
+        ingest_after_update,
+        &[
+            ("body:hello", Ok(1)),
+            ("body:world", Ok(1)),
+            ("title:salutations", Ok(1)),
+        ],
+    )
+    .await;
+}
+
+#[tokio::test]
+async fn test_update_doc_mapping_dynamic_to_strict() {
+    let index_id = "update-dynamic-to-strict";
     let original_doc_mappings = json!({
         "mode": "dynamic",
     });
@@ -518,7 +419,8 @@ async fn test_update_doc_mappings_add_on_dynamic() {
     let updated_doc_mappings = json!({
         "field_mappings": [
             {"name": "body", "type": "text"}
-        ]
+        ],
+        "mode": "strict",
     });
     let ingest_after_update = &[json!({"body": "world"})];
     validate_search_across_doc_mapping_updates(
@@ -533,19 +435,21 @@ async fn test_update_doc_mappings_add_on_dynamic() {
 }
 
 #[tokio::test]
-async fn test_update_doc_mappings_add_field() {
-    let index_id = "update-add-field";
+async fn test_update_doc_mapping_add_field_on_strict() {
+    let index_id = "update-add-field-on-strict";
     let original_doc_mappings = json!({
         "field_mappings": [
             {"name": "body", "type": "text"},
-        ]
+        ],
+        "mode": "strict",
     });
     let ingest_before_update = &[json!({"body": "hello"})];
     let updated_doc_mappings = json!({
         "field_mappings": [
             {"name": "body", "type": "text"},
             {"name": "title", "type": "text"},
-        ]
+        ],
+        "mode": "strict",
     });
     let ingest_after_update = &[json!({"body": "world", "title": "salutations"})];
     validate_search_across_doc_mapping_updates(
